fix(basedata): clean up store and converter types
- Remove duplicate ItemTemplate type from store package
- Add LookupItemTemplate and GetItemTemplate helpers
- Fix InventoryType conversion in basedata_converter

+++ pkg/summit/tools/data/basedata_converter.go
@@ -102,7 +102,7 @@ func (dc *Converter) ConvertItems() ([]*basedata.ItemTemplate, error) {
 			Entry:         table.Uint32(r, 0),
 			DisplayID:     table.Uint32(r, displayCol),
 			InventoryType: wow.InventoryType(table.Uint32(r, typeCol)),
-			SheatheType:   uint8(table.Uint32(r, sheatheCol)),
+			SheatheType:   table.Uint32(r, sheatheCol),
 		})
 	}
